Add files via upload
This is the first part of a project where I use Selenium and BeautifulSoup to scrape perfume product data (including product names, brand names, and reviews) from the Ulta website. The collected data is stored in a pandas DataFrame and can be saved as a CSV file for further analysis. This is just the beginning of a larger project focused on gathering and analyzing e-commerce data from various online stores.